requirements.txt
aiohttp==3.12.14
aiogram==3.22.0
python-dotenv==1.2.1